SoftapControllerTI: don't remove the STA role when starting AP
Only start the AP role (as we support multi-role).

Note: This patch hardcodes the channel of the started AP to 1. This
can be changed from the code for now.

Note2: This patch also hard-codes the MAC address of the AP, to avoid a
problem where both interfaces can't be set up.

Change-Id: I3dca6ca521fbbd40e29af56be0f8079a528fe9cb